requirements.txt
altgraph==0.17.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cmake==4.0.3
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
labelImg==1.8.6
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.146
ultralytics-thop==2.0.14
urllib3==2.4.0
wheel==0.45.1
